Refactor log stream validators
